deep-interest-network/requirements.txt
absl-py==0.10
astor==0.8.0
gast==0.4.0
google-pasta==0.2
grpcio==1.34.0
h5py==3.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.1.1
numpy==1.19.2
protobuf==3.15.0
six==1.15.0
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.5.3
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Werkzeug==0.15.4
wrapt==1.12.1
